+++ extensions/platform-actions-storage-db/src/main/resources/io/medatarun/actions/infra/db/v003__actions_renames.sql
@@ -1,0 +1,2 @@
+UPDATE action_audit_event SET action_group_key = "models" WHERE action_group_key = "model";
+UPDATE action_audit_event SET action_group_key = "tags" WHERE action_group_key = "tag"