+++ extensions/tags-core/src/main/kotlin/io/medatarun/tags/core/actions/TagActionProvider.kt
@@ -50,7 +50,7 @@ class TagActionProvider(
     }
 
     companion object {
-        const val ACTION_GROUP_KEY = "tag"
+        const val ACTION_GROUP_KEY = "tags"
     }
 }