+++ ui/src/components/business/model/RelationshipsTable.tsx
@@ -51,7 +51,7 @@ export function RelationshipsTable({
   const { t } = useAppI18n();
   const actionRegistry = useActionRegistry();
   const itemActions = actionRegistry.findActionDescriptors([
-    "model/relationship_delete",
+    "models/relationship_delete",
   ]);
   const { isDetailLevelTech } = useDetailLevelContext();
   const styles = useStyles();