+++ ui/src/components/business/model/model.storage.ts
@@ -54,7 +54,7 @@ export async function modelSearch(
     },
     fields: ["location"],
   };
-  return performer.executeJson<SearchResults>("model/search", payload);
+  return performer.executeJson<SearchResults>("models/search", payload);
 }
 export const useModelSearch = (req: ModelSearchReq) => {
   const { performer } = useActionPerformer();
@@ -70,7 +70,7 @@ export const useModelCompare = (req: ModelCompareReq | undefined) => {
     queryKey: ["model", "compare", req],
     queryFn: async () => {
       if (req === undefined) return undefined;
-      return performer.executeJson<ModelCompareDto>("model/model_compare", {
+      return performer.executeJson<ModelCompareDto>("models/model_compare", {
         leftModelRef: "id:" + req.leftModelId,
         leftModelVersion: req.leftModelVersion,
         rightModelRef: "id:" + req.rightModelId,
@@ -87,7 +87,7 @@ export function useModelHistoryVersions(modelId: string) {
     queryKey: ["model", modelId, "history_versions"],
     queryFn: () =>
       performer.executeJson<ModelChangeEventListWithVersionDto>(
-        "model/history_versions",
+        "models/history_versions",
         {
           modelRef: "id:" + modelId,
         },
@@ -111,7 +111,7 @@ export function useModelHistoryVersionChanges(
         payload.version = version;
       }
       return performer.executeJson<ModelChangeEventListDto>(
-        "model/history_version_changes",
+        "models/history_version_changes",
         payload,
       );
     },
@@ -122,7 +122,7 @@ export const useModelUpdateName = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; value: string }) =>
-      performer.executeJson("model/model_update_name", {
+      performer.executeJson("models/model_update_name", {
         modelRef: "id:" + props.modelId,
         value: props.value,
       }),
@@ -132,7 +132,7 @@ export const useModelUpdateDescription = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; value: string }) =>
-      performer.executeJson("model/model_update_description", {
+      performer.executeJson("models/model_update_description", {
         modelRef: "id:" + props.modelId,
         value: props.value,
       }),
@@ -142,7 +142,7 @@ export const useModelUpdateDocumentationHome = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; value: string }) =>
-      performer.executeJson("model/model_update_documentation_link", {
+      performer.executeJson("models/model_update_documentation_link", {
         modelRef: "id:" + props.modelId,
         value: props.value,
       }),
@@ -152,7 +152,7 @@ export const useModelUpdateKey = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; value: string }) =>
-      performer.executeJson("model/model_update_key", {
+      performer.executeJson("models/model_update_key", {
         modelRef: "id:" + props.modelId,
         value: props.value,
       }),
@@ -162,7 +162,7 @@ export const useModelAddTag = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; tag: string }) =>
-      performer.executeJson("model/model_add_tag", {
+      performer.executeJson("models/model_add_tag", {
         modelRef: "id:" + props.modelId,
         tag: props.tag,
       }),
@@ -172,7 +172,7 @@ export const useModelDeleteTag = () => {
   const { performer } = useActionPerformer();
   return useMutation({
     mutationFn: (props: { modelId: string; tag: string }) =>
-      performer.executeJson("model/model_delete_tag", {
+      performer.executeJson("models/model_delete_tag", {
         modelRef: "id:" + props.modelId,
         tag: props.tag,
       }),
@@ -194,13 +194,13 @@ function useTypeMutation<P extends ActionPayload>(actionRef: ActionKey) {
 }
 
 export const useTypeUpdateName = () => {
-  return useTypeMutation<{ value: string }>("model/type_update_name");
+  return useTypeMutation<{ value: string }>("models/type_update_name");
 };
 export const useTypeUpdateKey = () => {
-  return useTypeMutation<{ value: string }>("model/type_update_key");
+  return useTypeMutation<{ value: string }>("models/type_update_key");
 };
 export const useTypeUpdateDescription = () => {
-  return useTypeMutation<{ value: string }>("model/type_update_description");
+  return useTypeMutation<{ value: string }>("models/type_update_description");
 };
 
 function useEntityMutation<P extends ActionPayload>(actionRef: ActionKey) {
@@ -218,26 +218,26 @@ function useEntityMutation<P extends ActionPayload>(actionRef: ActionKey) {
 }
 
 export const useEntityUpdateName = () => {
-  return useEntityMutation<{ value: string }>("model/entity_update_name");
+  return useEntityMutation<{ value: string }>("models/entity_update_name");
 };
 export const useEntityUpdateKey = () => {
-  return useEntityMutation<{ value: string }>("model/entity_update_key");
+  return useEntityMutation<{ value: string }>("models/entity_update_key");
 };
 export const useEntityUpdateDescription = () => {
   return useEntityMutation<{ value: string }>(
-    "model/entity_update_description",
+    "models/entity_update_description",
   );
 };
 export const useEntityUpdateDocumentationHome = () => {
   return useEntityMutation<{ value: string }>(
-    "model/entity_update_documentation_link",
+    "models/entity_update_documentation_link",
   );
 };
 export const useEntityAddTag = () => {
-  return useEntityMutation<{ tag: string }>("model/entity_add_tag");
+  return useEntityMutation<{ tag: string }>("models/entity_add_tag");
 };
 export const useEntityDeleteTag = () => {
-  return useEntityMutation<{ tag: string }>("model/entity_delete_tag");
+  return useEntityMutation<{ tag: string }>("models/entity_delete_tag");
 };
 
 function useEntityAttributeMutation<P extends ActionPayload>(
@@ -261,37 +261,37 @@ function useEntityAttributeMutation<P extends ActionPayload>(
 
 export const useEntityAttributeUpdateName = () => {
   return useEntityAttributeMutation<{ value: string }>(
-    "model/entity_attribute_update_name",
+    "models/entity_attribute_update_name",
   );
 };
 export const useEntityAttributeUpdateDescription = () => {
   return useEntityAttributeMutation<{ value: string }>(
-    "model/entity_attribute_update_description",
+    "models/entity_attribute_update_description",
   );
 };
 export const useEntityAttributeUpdateKey = () => {
   return useEntityAttributeMutation<{ value: string }>(
-    "model/entity_attribute_update_key",
+    "models/entity_attribute_update_key",
   );
 };
 export const useEntityAttributeUpdateType = () => {
   return useEntityAttributeMutation<{ value: string }>(
-    "model/entity_attribute_update_type",
+    "models/entity_attribute_update_type",
   );
 };
 export const useEntityAttributeUpdateOptional = () => {
   return useEntityAttributeMutation<{ value: boolean }>(
-    "model/entity_attribute_update_optional",
+    "models/entity_attribute_update_optional",
   );
 };
 export const useEntityAttributeAddTag = () => {
   return useEntityAttributeMutation<{ tag: string }>(
-    "model/entity_attribute_add_tag",
+    "models/entity_attribute_add_tag",
   );
 };
 export const useEntityAttributeDeleteTag = () => {
   return useEntityAttributeMutation<{ tag: string }>(
-    "model/entity_attribute_delete_tag",
+    "models/entity_attribute_delete_tag",
   );
 };
 
@@ -313,25 +313,27 @@ function useRelationshipMutation<P extends ActionPayload>(
 
 export const useRelationshipUpdateName = () => {
   return useRelationshipMutation<{ value: string }>(
-    "model/relationship_update_name",
+    "models/relationship_update_name",
   );
 };
 export const useRelationshipUpdateDescription = () => {
   return useRelationshipMutation<{ value: string }>(
-    "model/relationship_update_description",
+    "models/relationship_update_description",
   );
 };
 export const useRelationshipUpdateKey = () => {
   return useRelationshipMutation<{ value: string }>(
-    "model/relationship_update_key",
+    "models/relationship_update_key",
   );
 };
 export const useRelationshipAddTag = () => {
-  return useRelationshipMutation<{ tag: string }>("model/relationship_add_tag");
+  return useRelationshipMutation<{ tag: string }>(
+    "models/relationship_add_tag",
+  );
 };
 export const useRelationshipDeleteTag = () => {
   return useRelationshipMutation<{ tag: string }>(
-    "model/relationship_delete_tag",
+    "models/relationship_delete_tag",
   );
 };
 
@@ -360,37 +362,37 @@ function useRelationshipAttributeMutation<P extends ActionPayload>(
 
 export const useRelationshipAttributeUpdateName = () => {
   return useRelationshipAttributeMutation<{ value: string }>(
-    "model/relationship_attribute_update_name",
+    "models/relationship_attribute_update_name",
   );
 };
 export const useRelationshipAttributeUpdateDescription = () => {
   return useRelationshipAttributeMutation<{ value: string }>(
-    "model/relationship_attribute_update_description",
+    "models/relationship_attribute_update_description",
   );
 };
 export const useRelationshipAttributeUpdateKey = () => {
   return useRelationshipAttributeMutation<{ value: string }>(
-    "model/relationship_attribute_update_key",
+    "models/relationship_attribute_update_key",
   );
 };
 export const useRelationshipAttributeUpdateType = () => {
   return useRelationshipAttributeMutation<{ value: string }>(
-    "model/relationship_attribute_update_type",
+    "models/relationship_attribute_update_type",
   );
 };
 export const useRelationshipAttributeUpdateOptional = () => {
   return useRelationshipAttributeMutation<{ value: boolean }>(
-    "model/relationship_attribute_update_optional",
+    "models/relationship_attribute_update_optional",
   );
 };
 export const useRelationshipAttributeAddTag = () => {
   return useRelationshipAttributeMutation<{ tag: string }>(
-    "model/relationship_attribute_add_tag",
+    "models/relationship_attribute_add_tag",
   );
 };
 export const useRelationshipAttributeDeleteTag = () => {
   return useRelationshipAttributeMutation<{ tag: string }>(
-    "model/relationship_attribute_delete_tag",
+    "models/relationship_attribute_delete_tag",
   );
 };
 
@@ -402,7 +404,7 @@ export function useModelList() {
   const { performer } = useActionPerformer();
   return useQuery({
     queryKey: ["model", "list"],
-    queryFn: () => performer.executeJson<ModelListDto>("model/model_list", {}),
+    queryFn: () => performer.executeJson<ModelListDto>("models/model_list", {}),
   });
 }