+++ ui/src/views/attribute/AttributeEditPage.tsx
@@ -171,8 +171,8 @@ export function AttributeView({
 
   const actions = actionRegistry.findActionDescriptors(
     parentType == "entity"
-      ? ["model/entity_attribute_delete"]
-      : ["model/relationship_attribute_delete"],
+      ? ["models/entity_attribute_delete"]
+      : ["models/relationship_attribute_delete"],
   );
 
   const handleClickModel = () => {
@@ -203,12 +203,12 @@ export function AttributeView({
     parentType === "relationship" ? (parent as RelationshipDto) : null;
 
   const updateNameDisabled = parentAsEntity
-    ? !sec.canExecuteAction("model/entity_attribute_update_name")
-    : !sec.canExecuteAction("model/relationship_attribute_update_name");
+    ? !sec.canExecuteAction("models/entity_attribute_update_name")
+    : !sec.canExecuteAction("models/relationship_attribute_update_name");
 
   const updateDescriptionDisabled = parentAsEntity
-    ? !sec.canExecuteAction("model/entity_attribute_update_description")
-    : !sec.canExecuteAction("model/relationship_attribute_update_description");
+    ? !sec.canExecuteAction("models/entity_attribute_update_description")
+    : !sec.canExecuteAction("models/relationship_attribute_update_description");
 
   const displayedSubject: ActionDisplayedSubject =
     parentAsEntity !== null
@@ -407,26 +407,26 @@ export function AttributeOverview({
     model.isEntityAttributePK(parentAsEntity.id, attribute.id);
 
   const updateKeyDisabled = parentAsEntity
-    ? !sec.canExecuteAction("model/entity_attribute_update_key")
-    : !sec.canExecuteAction("model/relationship_attribute_update_key");
+    ? !sec.canExecuteAction("models/entity_attribute_update_key")
+    : !sec.canExecuteAction("models/relationship_attribute_update_key");
 
   const updateTypeDisabled = parentAsEntity
-    ? !sec.canExecuteAction("model/entity_attribute_update_type")
-    : !sec.canExecuteAction("model/relationship_attribute_update_type");
+    ? !sec.canExecuteAction("models/entity_attribute_update_type")
+    : !sec.canExecuteAction("models/relationship_attribute_update_type");
 
   const updateTagDisabled = parentAsEntity
     ? !sec.canExecuteActions(
-        "model/entity_attribute_add_tag",
-        "model/entity_attribute_delete_tag",
+        "models/entity_attribute_add_tag",
+        "models/entity_attribute_delete_tag",
       )
     : !sec.canExecuteActions(
-        "model/relationship_attribute_add_tag",
-        "model/relationship_attribute_delete_tag",
+        "models/relationship_attribute_add_tag",
+        "models/relationship_attribute_delete_tag",
       );
 
   const updateOptionalDisabled = parentAsEntity
-    ? !sec.canExecuteAction("model/entity_attribute_update_optional")
-    : !sec.canExecuteAction("model/relationship_attribute_update_optional");
+    ? !sec.canExecuteAction("models/entity_attribute_update_optional")
+    : !sec.canExecuteAction("models/relationship_attribute_update_optional");
 
   const handleChangeKey = (value: string) => {
     if (parentAsEntity) {